+++ src/app/bug-buster/page.tsx
@@ -1,36 +1,10 @@
 "use client";
 //import BugBusterList from "@/components/BugBusterList/BugBusterList";
 import BugBusterList from "@/components/BugBusterList/BugBusterList";
-import { fetchBugBusterList } from "@/data/fetchBugBuster";
-import { IBugBuster, } from "@/definition/definition";
-//import dynamic from "next/dynamic";
-import {useEffect, useState } from "react";
+
 
 export default  function BugBusterPage() {
-  const [bugBusterList, setBugBusterList] = useState<IBugBuster[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-
-  console.log(bugBusterList);
-
-  useEffect(() => {
-    async function fetchBugList(){
-      setIsLoading(false)
-      try{
-        setIsLoading(true);
-        const data = await fetchBugBusterList();
-        const bugList = data.bugList
-        setBugBusterList(bugList);
-
-      }catch(error){
-        if(error instanceof Error){
-          return "Server Error";
-        }
-      }finally{
-        setIsLoading(false);
-      }
-    }
-    fetchBugList();
-  }, [])
+  
   return (
     <>
       <section className="py-16 px-4">
@@ -42,9 +16,7 @@ export default  function BugBusterPage() {
             </p>
           </div>
         </header>
-       {
-        isLoading ? <p>Loading</p> : <BugBusterList bugBusterList={bugBusterList}/>
-       }
+        <BugBusterList />
       </section>
     </>
   );
