+++ src/components/BugBusterList/BugBusterList.tsx
@@ -2,14 +2,36 @@
 
 import { IBugBuster } from "@/definition/definition";
 import Link from "next/link";
+import { fetchBugBusterList } from "@/data/fetchBugBuster";
+import {useEffect, useState } from "react";
 
-interface BugBusterListProps {
-  bugBusterList: IBugBuster[];
-}
-export default function BugBusterList({ bugBusterList }: BugBusterListProps) {
+export default function BugBusterList() {
+
+  const [bugBusterList, setBugBusterList] = useState<IBugBuster[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    async function fetchBugList(){
+      setIsLoading(false)
+      try{
+        setIsLoading(true);
+        const data = await fetchBugBusterList();
+        const bugList = data.bugList
+        setBugBusterList(bugList);
+
+      }catch(error){
+        if(error instanceof Error){
+          return "Server Error";
+        }
+      }finally{
+        setIsLoading(false);
+      }
+    }
+    fetchBugList();
+  }, [])
   return (
     <>
-      <div>
+      { isLoading ? <p>Loading...</p> : <div>
         {bugBusterList.map((bugBuster) => (
           <Link
             key={bugBuster.id}
@@ -20,6 +42,7 @@ export default function BugBusterList({ bugBusterList }: BugBusterListProps) {
           </Link>
         ))}
       </div>
+}
     </>
   );
 }
